Moved things around to satisfy linter
